test: move recorder seeking to transport spec
